requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
pymongo==4.5.0
python-dotenv==1.0.0
werkzeug==2.3.7
WTForms==3.0.1
email-validator==2.0.0
markupsafe==2.1.3
itsdangerous==2.1.2
